backend/requirements.txt
Django>=4.2
gunicorn
psycopg2-binary
djangorestframework
django-cors-headers
dj-database-url
django-environ
requests
djangorestframework
